+++ src/Plugins/Api.WebApi/Controllers/OData/ProductsController.cs
@@ -12,6 +12,9 @@ using SmartStore.Services.Seo;
 using SmartStore.Web.Framework.WebApi;
 using SmartStore.Web.Framework.WebApi.OData;
 using SmartStore.Web.Framework.WebApi.Security;
+using System.Web.Http.OData;
+using SmartStore.Plugin.Api.WebApi.Services;
+using System.Web.Http.OData.Query;
 
 namespace SmartStore.Plugin.Api.WebApi.Controllers.OData
 {
@@ -21,15 +24,18 @@ namespace SmartStore.Plugin.Api.WebApi.Controllers.OData
 		private readonly Lazy<IWorkContext> _workContext;
 		private readonly Lazy<IPriceCalculationService> _priceCalculationService;
 		private readonly Lazy<IUrlRecordService> _urlRecordService;
+		private readonly Lazy<IProductAttributeService> _productAttributeService;
 
 		public ProductsController(
 			Lazy<IWorkContext> workContext,
 			Lazy<IPriceCalculationService> priceCalculationService,
-			Lazy<IUrlRecordService> urlRecordService)
+			Lazy<IUrlRecordService> urlRecordService,
+			Lazy<IProductAttributeService> productAttributeService)
 		{
 			_workContext = workContext;
 			_priceCalculationService = priceCalculationService;
 			_urlRecordService = urlRecordService;
+			_productAttributeService = productAttributeService;
 		}
 
 		protected override IQueryable<Product> GetEntitySet()
@@ -217,6 +223,125 @@ namespace SmartStore.Plugin.Api.WebApi.Controllers.OData
 		{
 			return CalculatePrice(key, true);
 		}
+
+		[HttpPost, WebApiQueryable(PagingOptional = true)]
+		public IQueryable<ProductVariantAttributeCombination> CreateAttributeCombinations(int key)
+		{
+			var entity = GetEntityByKeyNotNull(key);
+
+			this.ProcessEntity(() =>
+			{
+				_productAttributeService.Value.CreateAllProductVariantAttributeCombinations(entity);
+				return null;
+			});
+
+			return entity.ProductVariantAttributeCombinations.AsQueryable();
+		}
+
+		[HttpPost, WebApiQueryable(PagingOptional = true)]
+		public IQueryable<ProductVariantAttribute> ManageAttributes(int key, ODataActionParameters parameters)
+		{
+			var entity = GetExpandedEntity<ICollection<ProductVariantAttribute>>(key, x => x.ProductVariantAttributes);
+			var result = new List<ProductVariantAttributeValue>();
+
+			this.ProcessEntity(() =>
+			{
+				bool synchronize = parameters.GetValue<string, bool>("Synchronize");
+				var data = (parameters["Attributes"] as IEnumerable<ManageAttributeType>).Where(x => x.Name.HasValue()).ToList();
+
+				var allAttributes = _productAttributeService.Value.GetAllProductAttributes();
+
+				foreach (var srcAttr in data)
+				{
+					var productAttribute = allAttributes.FirstOrDefault(x => x.Name.IsCaseInsensitiveEqual(srcAttr.Name));
+
+					if (productAttribute == null)
+					{
+						productAttribute = new ProductAttribute() { Name = srcAttr.Name };
+						_productAttributeService.Value.InsertProductAttribute(productAttribute);
+					}
+
+					var attribute = entity.ProductVariantAttributes.FirstOrDefault(x => x.ProductAttribute.Name.IsCaseInsensitiveEqual(srcAttr.Name));
+
+					if (attribute == null)
+					{
+						attribute = new ProductVariantAttribute()
+						{
+							ProductId = entity.Id,
+							ProductAttributeId = productAttribute.Id,
+							AttributeControlTypeId = srcAttr.ControlTypeId,
+							DisplayOrder = entity.ProductVariantAttributes.OrderByDescending(x => x.DisplayOrder).Select(x => x.DisplayOrder).FirstOrDefault() + 1,
+							IsRequired = srcAttr.IsRequired
+						};
+
+						entity.ProductVariantAttributes.Add(attribute);
+						Service.UpdateProduct(entity);
+					}
+					else if (synchronize)
+					{
+						if (srcAttr.Values.Count <= 0)
+						{
+							_productAttributeService.Value.DeleteProductVariantAttribute(attribute);
+						}
+						else
+						{
+							attribute.AttributeControlTypeId = srcAttr.ControlTypeId;
+							attribute.IsRequired = srcAttr.IsRequired;
+
+							Service.UpdateProduct(entity);
+						}
+					}
+
+					int maxDisplayOrder = attribute.ProductVariantAttributeValues.OrderByDescending(x => x.DisplayOrder).Select(x => x.DisplayOrder).FirstOrDefault();
+
+					foreach (var srcVal in srcAttr.Values.Where(x => x.Name.HasValue()))
+					{
+						var value = attribute.ProductVariantAttributeValues.FirstOrDefault(x => x.Name.IsCaseInsensitiveEqual(srcVal.Name));
+
+						if (value == null)
+						{
+							value = new ProductVariantAttributeValue()
+							{
+								ProductVariantAttributeId = attribute.Id,
+								Name = srcVal.Name,
+								Alias = srcVal.Alias,
+								ColorSquaresRgb = srcVal.ColorSquaresRgb,
+								PriceAdjustment = srcVal.PriceAdjustment,
+								WeightAdjustment = srcVal.WeightAdjustment,
+								IsPreSelected = srcVal.IsPreSelected,
+								DisplayOrder = ++maxDisplayOrder
+							};
+
+							attribute.ProductVariantAttributeValues.Add(value);
+							Service.UpdateProduct(entity);
+						}
+						else if (synchronize)
+						{
+							value.Alias = srcVal.Alias;
+							value.ColorSquaresRgb = srcVal.ColorSquaresRgb;
+							value.PriceAdjustment = srcVal.PriceAdjustment;
+							value.WeightAdjustment = srcVal.WeightAdjustment;
+							value.IsPreSelected = srcVal.IsPreSelected;
+
+							Service.UpdateProduct(entity);
+						}
+					}
+
+					if (synchronize)
+					{
+						foreach (var dstVal in attribute.ProductVariantAttributeValues.ToList())
+						{
+							if (!srcAttr.Values.Exists(x => x.Name.IsCaseInsensitiveEqual(dstVal.Name)))
+								_productAttributeService.Value.DeleteProductVariantAttributeValue(dstVal);
+						}
+					}
+				}
+				return null;
+			});
+
+			return entity.ProductVariantAttributes.AsQueryable();
+		}
+
 
 		//[HttpGet, WebApiQueryable]
 		//public IQueryable<RelatedProduct> GetRelatedProducts(int key)