+++ changelog.md
@@ -12,6 +12,7 @@
 * (Developer) Implemented _IWidgetProvider_. Allows request scoped registration of action routes to be injectable into widget zones. Perfect for custom action filters.
 * (Developer) Simple widgets: the model of the parent action view context now gets passed to a widget.
 * (Developer) New IoC method ContainerManager.InjectProperties()
+* #393 Web API: Implement OData actions for simpler working with product attributes
 
 ###Improvements###
 * Task Scheduler: